free_list should remove items from the list
free_list would only free the memory of the items and not removing
them from the list, i.e. lefting dangling pointers.

We can fix it changing the loop.  We could also drop the check in
configure.ac for SIMPLEQ_FOREACH_SAFE eventually since we're not using
it anymore, but I'll leave it in place for a bit still.